refactor: moving types back to feature needed in
