+++ ui/src/lib/features/k8s/daemonsets/store.ts
@@ -3,7 +3,8 @@
 
 import type { V1DaemonSet as Resource } from '@kubernetes/client-node'
 
-import { type CommonRow, type ResourceStoreInterface, type ResourceWithTable } from '$lib/types'
+import { type CommonRow } from '$lib/types'
+import { type ResourceStoreInterface, type ResourceWithTable } from '$lib/features/k8s/types'
 import { ResourceStore } from '../store'
 import { type ColumnWrapper } from '../types'
