+++ ui/src/lib/features/k8s/types.ts
@@ -1,1 +1,38 @@
+import type { KubernetesObject } from '@kubernetes/client-node'
+import { type Writable } from 'svelte/store'
+
+import { type CommonRow } from '../../types'
+
 export type ColumnWrapper<T> = [name: keyof T, styles?: string][]
+
+export enum SearchByType {
+  ANYWHERE = 'Anywhere',
+  METADATA = 'Metadata',
+  NAME = 'Name',
+}
+
+export interface ResourceWithTable<T extends KubernetesObject, U extends CommonRow> {
+  resource: T
+  table: U
+}
+
+export interface ResourceStoreInterface<T extends KubernetesObject, U extends CommonRow> {
+  // Start the EventSource and update the resources
+  start: () => () => void
+  // Sort the table by the key
+  sortByKey: (key: keyof U) => void
+  // Store for search text
+  search: Writable<string>
+  // Store for search by type
+  searchBy: Writable<SearchByType>
+  // Store for sortBy key
+  sortBy: Writable<keyof U>
+  // Store for sort direction
+  sortAsc: Writable<boolean>
+  // The list of search types
+  searchTypes: SearchByType[]
+  // Subscribe to the filtered and sorted resources
+  subscribe: (run: (value: ResourceWithTable<T, U>[]) => void) => () => void
+  // Store for namespace
+  namespace: Writable<string>
+}
