+++ ui/src/lib/types.ts
@@ -1,6 +1,3 @@
-import type { KubernetesObject } from '@kubernetes/client-node'
-import { type Writable } from 'svelte/store'
-
 export interface CommonRow {
   name: string
   namespace?: string
@@ -19,36 +16,4 @@ export type PeprEvent = {
   repeated?: number
   ts?: string
   epoch: number
-}
-
-export enum SearchByType {
-  ANYWHERE = 'Anywhere',
-  METADATA = 'Metadata',
-  NAME = 'Name',
-}
-
-export interface ResourceWithTable<T extends KubernetesObject, U extends CommonRow> {
-  resource: T
-  table: U
-}
-
-export interface ResourceStoreInterface<T extends KubernetesObject, U extends CommonRow> {
-  // Start the EventSource and update the resources
-  start: () => () => void
-  // Sort the table by the key
-  sortByKey: (key: keyof U) => void
-  // Store for search text
-  search: Writable<string>
-  // Store for search by type
-  searchBy: Writable<SearchByType>
-  // Store for sortBy key
-  sortBy: Writable<keyof U>
-  // Store for sort direction
-  sortAsc: Writable<boolean>
-  // The list of search types
-  searchTypes: SearchByType[]
-  // Subscribe to the filtered and sorted resources
-  subscribe: (run: (value: ResourceWithTable<T, U>[]) => void) => () => void
-  // Store for namespace
-  namespace: Writable<string>
 }
